+++ src/game/Shot.java
@@ -31,7 +31,7 @@ public class Shot extends Shape {
      * @param vy
      * @param radius
      */
-    public Shot(int x, int y, int vx, int vy, int radius) {
+    public Shot(double x, double y, double vx, double vy, int radius) {
         super(x, y, vx, vy);
         this.radius = radius;        
     }
@@ -46,6 +46,6 @@ public class Shot extends Shape {
     @Override
     public void paint(Graphics2D g2d) {
         int diameter = 2 * radius;
-        g2d.fillOval(x - radius, y - radius, diameter, diameter);
+        g2d.fillOval((int)x - radius, (int)y - radius, diameter, diameter);
     }
 }
